+++ app/api/common.ts
@@ -1,2 +1,4 @@
 /** Root path for the API */
 export const API_BASE = '/api'
+
+export const URL_VIDEO_FILES = `${API_BASE}/video/files`
