+++ app/api/video_client.ts
@@ -1,0 +1,7 @@
+import { Dir } from '../../lib'
+
+import * as common from './common'
+
+export async function fetch_files() {
+	return fetch(common.URL_VIDEO_FILES).then((x) => x.json() as Promise<Dir>)
+}
