+++ lib/index.ts
@@ -1,0 +1,6 @@
+/**
+ * @file This library contains shared definitions between the frontend and
+ * backend.
+ */
+
+export * from './entities'
