--- lib/video_types.ts
@@ -1,14 +1,0 @@
-export type DirEntry = File | Dir
-
-export type File = {
-	type: 'video' | 'subtitle'
-	name: string
-	path: string
-}
-
-export type Dir = {
-	type: 'dir'
-	name: string
-	path: string
-	list: DirEntry[]
-}
